Modified CLI and renamed packages.
