+++ dspace-api/src/test/resources/org/dspace/iiif/canvasdimension/cat.jpg
