+++ dspace/config/dspace.cfg
@@ -733,7 +733,7 @@ event.consumer.authority.class = org.dspace.authority.indexer.AuthorityConsumer
 event.consumer.authority.filters = Item+Modify|Modify_Metadata
 
 # iiif consumer
-event.consumer.iiif.class = org.dspace.iiif.IIIFCacheEventConsumer
+event.consumer.iiif.class = org.dspace.iiif.consumer.IIIFCacheEventConsumer
 event.consumer.iiif.filters = Item+Modify:Item+Modify_Metadata:Item+Delete:Item+Remove:Bundle+ALL:Bitstream+All
 
 # ...set to true to enable testConsumer messages to standard output
